==> 0-themes.R <==
theme_simple <- function(s){
  theme(panel.grid = element_blank(),
        panel.background = element_blank(),
        axis.line = element_line(color = "grey"),
        axis.text = element_text(size = 12))
}
